scripts for uninstall bundled edge

+++ PowerShellProfile.ps1
@@ -188,32 +188,40 @@ function Update-Kodi {
 }
 
 function Uninstall-UselessPackages {
-    sudo Uninstall-UselessPackagesInternal
-    
     winget uninstall Cortana
     winget uninstall Clipchamp
-
 }
 
-function Uninstall-UselessPackagesInternal {
-    sudo Unregister-ScheduledTask *Edge* -Confirm:False
-    Get-ChildItem "C:\Program Files (x86)\Microsoft\EdgeWebView\*\Installer" -Recurse | Foreach-Object {
-        Push-Location $_
-        Copy-Item C:\SHELL\EDGEUNINSTALLER setup.exe
-        ./setup.exe --uninstall --force-uninstall --system-level --msedgewebview
-        Pop-Location
+function Uninstall-BundledEdge {
+    if (([System.Security.Principal.WindowsIdentity]::GetCurrent()).Groups -match "S-1-5-32-544") {
+        Unregister-ScheduledTask *Edge* -Confirm:$false
+
+        sc delete edgeupdate
+        sc delete edgeupdatem
+
+        taskkill /f /im MicrosoftEdgeUpdate.exe
+
+        Get-ChildItem "C:\Program Files (x86)\Microsoft\EdgeWebView\*\Installer" -Recurse | Foreach-Object {
+            Push-Location $_
+            Copy-Item C:\SHELL\EDGEUNINSTALLER setup.exe
+            ./setup.exe --uninstall --force-uninstall --system-level --msedgewebview
+            Pop-Location
+        }
+
+        Get-ChildItem "C:\Program Files (x86)\Microsoft\Edge\*\Installer" -Recurse | Foreach-Object {
+            Push-Location $_
+            Copy-Item C:\SHELL\EDGEUNINSTALLER setup.exe
+            ./setup.exe --uninstall --force-uninstall --system-level --msedge
+            Pop-Location
+        }
+
+        Remove-Item -Recurse -Force "C:\Program Files (x86)\Microsoft"
+
+        Remove-Item "HKCU:\Software\Microsoft\Windows NT\CurrentVersion\Fonts" -Force -Recurse
     }
-
-    Get-ChildItem "C:\Program Files (x86)\Microsoft\Edge\*\Installer" -Recurse | Foreach-Object {
-        Push-Location $_
-        Copy-Item C:\SHELL\EDGEUNINSTALLER setup.exe
-        ./setup.exe --uninstall --force-uninstall --system-level --msedge
-        Pop-Location
+    else {
+        Write-Output "Run as Administrator and try again"
     }
-
-    Remove-Item -Recurse -Force "C:\Program Files (x86)\Microsoft"
-
-    Remove-Item "HKCU:\Software\Microsoft\Windows NT\CurrentVersion\Fonts" -Force -Recurse
 }
 
 function Start-AngryBirds {
